requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==25.1.0
beautifulsoup4==4.13.3
bech32==1.2.0
black==24.10.0
CacheControl==0.14.3
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coincurve==21.0.0
cryptography==45.0.6
curlify==2.2.1
dataclasses-json==0.6.7
distro==1.9.0
Django==5.1.4
django-admin-sortable2==2.2.4
django-ckeditor-5==0.2.17
django-cors-headers==4.6.0
django-crontab==0.7.1
django-debug-toolbar==4.4.6
django-filter==25.1
django-rosetta==0.10.1
django-solo==2.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-standardized-errors==0.14.1
facebook_business==21.0.5
fastecdsa==3.0.0
firebase_admin==7.1.0
frozenlist==1.5.0
geoip2==5.1.0
google-api-core==2.25.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.3.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.74.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.14.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
hyperframe==6.1.0
idna==3.10
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-openai==0.3.27
langchain-tavily==0.2.7
langchain-text-splitters==0.3.8
langgraph==0.5.2
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.72
langsmith==0.4.5
logging==0.4.9.6
mailjet-rest==1.3.4
Markdown==3.7
marshmallow==3.26.1
maxminddb==2.7.0
msgpack==1.1.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.2.1
oauthlib==3.2.2
openai==1.95.1
orjson==3.10.15
ormsgpack==1.10.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
pip-autoremove==0.10.0
platformdirs==4.3.6
polib==1.2.0
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic-settings==2.10.1
pydantic_core==2.27.2
PyJWT==2.10.1
pyparsing==3.2.1
python-decouple==3.8
python-dotenv==1.1.1
python-monkey-business==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
sqlparse==0.5.3
stripe==11.4.1
tavily-python==0.7.9
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
wheel==0.45.1
xxhash==3.5.0
yarl==1.18.3
zstandard==0.23.0
